requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9

openai==1.51.0
httpx==0.27.2

oracledb==2.4.1

apscheduler==3.10.4

python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2

aiofiles==24.1.0
